requirements.txt
boto3>=1.34.0
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
requests>=2.31.0
python-dateutil>=2.8.2
jinja2>=3.1.2
python-multipart>=0.0.6
aiofiles>=23.2.1

# Development tools
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
autopep8>=2.0.0
